add end-to-end test that uses QUIC
This polynomial evaluation test gets one field element as input per client and
evaluates a fixed multi-variate polynomial with each input as a variable.
There is no optimization at all, so intermediate results are not cached and
x^4 is evaluated as x*x*x*x, not x^2*x^2.
The QUIC network is used.